requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
openai
tiktoken
faiss-cpu
huggingface_hub
pytesseract
pillow
python-docx
pandas
textract
openpyxl
python-pptx
tensorflow
os
